backend/app/requirements.txt
fastapi
uvicorn
qdrant-client
pillow
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
watchdog
python-multipart
python-dotenv